requirements.txt
python-dotenv
requests
flask
schedule
yt-dlp
